st-gpt4o/requirement.txt
requests  
streamlit  
json  
ast  
openai==0.27.8
python-dotenv  
Pillow  
numpy  
